* commit '018ddfa723dfeacc8b0349f634663f712db46cef':
  merchant bug fixes

+++ api/models/Merchants.js
@@ -60,7 +60,7 @@ module.exports = {
             }
         },
         fees: {
-            type: Sequelize.INTEGER(3),
+            type: Sequelize.FLOAT,
             required: true,
             allowNull: false,
             validate: {
